+++ packages/components/organisms/f-footer/test/specs/component/f-footer.component.desktop.spec.js
@@ -51,7 +51,7 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
     });
 
     forEach(['gb', 'au', 'ie', 'nz', 'dk', 'es', 'it', 'no'])
-    .it('should not show courier links and country selector when options are unselected', expectedLocale => {
+    .it('should not show courier links and country selector for country code "%s" when options are unselected', expectedLocale => {
         // Arrange
         const footerData = {
             locale: expectedLocale,
@@ -69,7 +69,7 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
     });
 
     forEach(['au', 'ie', 'nz'])
-    .it('should show courier links when option is selected', expectedLocale => {
+    .it('should show courier links for country code "%s" when option is selected', expectedLocale => {
         // Arrange
         const footerData = {
             locale: expectedLocale,
@@ -86,7 +86,7 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
     });
 
     forEach(['gb', 'es', 'it', 'no'])
-    .it('should never show courier links, even when option is selected', () => {
+    .it('should never show courier links for country code "%s", even when option is selected', expectedLocale => {
         // Arrange
         const footerData = {
             locale: expectedLocale,
@@ -103,7 +103,7 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
     });
 
     forEach(['au', 'ie', 'nz', 'dk', 'es', 'it', 'no'])
-    .it('should always show country selector when selected', expectedLocale => {
+    .it('should always show country selector for country code "%s" when selected', expectedLocale => {
         // Arrange
         const footerData = {
             locale: expectedLocale,
@@ -119,8 +119,8 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
         expect(footer.isCountrySelectorDisplayed()).toBe(true);
     });
 
-    forEach(['gb', 'au', 'ie', 'nz', 'dk', 'es', 'it', 'no'])
-    .it('should display the corresponding icon for each locale selected', expectedLocale => {
+    forEach(['au', 'gb', 'nz', 'ie', 'dk', 'es', 'it'])
+    .it('should display the corresponding country code ("%s") icon for each locale selected', expectedLocale => {
         // Arrange
         const footerData = {
             locale: expectedLocale,
@@ -136,9 +136,46 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
         expect(footer.isCurrentCountryIconDisplayed(expectedLocale)).toBe(true);
     });
 
-    forEach([['au', 'au'], ['at', 'at'], ['be', 'be-en'], ['bg', 'bg'], ['ca_en', 'skipthedishes.com'], ['ca_fr', 'skipthedishes.com/fr'], ['dk', '.dk'], ['jet_fr', '.fr'], ['de', '.de'], ['ie', '.ie'], ['il', '.il'], ['it', '.it'],
-        ['lu', 'lu-en'], ['nl', '.nl'], ['nz', '.nz'], ['no', '.no'], ['pl', '.pl'], ['pt', '/pt'], ['ro', '/ro'], ['es', '.es'], ['ch_ch', '.ch'], ['ch_en', '/en'], ['ch_fr', '/fr']])
-    .it('should display all countries and redirect to correct URL', (country, expectedUrl) => {
+    forEach([
+        ['au', 'au'],
+        ['at', 'at'],
+        ['no', 'no'],
+        ['dk', 'dk'],
+        ['be', 'be-en'],
+        ['bg', 'bg'],
+        ['ca_en', 'skipthedishes.com'],
+        ['jet_fr', '.fr'],
+        ['de', '.de'],
+        ['ie', '.ie'],
+        ['il', '.il'],
+        ['it', '.it']
+    ]).it('should display link for country code "%s" and redirect to correct URL', (country, expectedUrl) => { 
+        // Act
+        footer.clickCountrySelectorButton();
+        footer.expectedCountry = country;
+
+        // Assert
+        expect(footer.isCountryLinkItemDisplayed()).toBe(true);
+
+        // Act
+        footer.clickCountryLinkItem();
+
+        // Assert
+        expect(browser.getUrl()).toContain(expectedUrl);
+    });
+
+    forEach([
+        ['lu', 'lu-en'],
+        ['nl', '.nl'],
+        ['nz', '.nz'],
+        ['pl', '.pl'],
+        ['pt', '/pt'],
+        ['ro', '/ro'],
+        ['es', '.es'],
+        ['ch_ch', '.ch'],
+        ['ch_en', '/en'],
+        ['ch_fr', '/fr']])
+    .it('should display link for country code "%s" and redirect to correct URL', (country, expectedUrl) => {
         // Act
         footer.clickCountrySelectorButton();
         footer.expectedCountry = country;